TruncatedNormal Fix and Cleanup (#48)
* fixed trunc normal; mandatory mu instead of dim

* minor fix

* minor fix

* minor fix

* clean up tests
